+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Prefix the change with one of these keywords:
 - Fixed: for any bug fixes.
 - Security: in case of vulnerabilities.
 
-## UNRELEASED
+## [0.6.0]
 - Removed: **BREAKING** unused props in several components 
 - Removed: **BREAKING** `onInitLayers` and unused `isOpen` props on draw tool
 - Added: `onEndInitialItems` prop for draw tool